requirements.txt
asgiref==3.2.5
astroid==2.4.2
boto3==1.12.31
botocore==1.15.31
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.4
django-cors-headers==3.2.1
django-crum==0.7.7
django-email-verification==0.0.2
django-environ==0.4.5
django-grappelli==2.14.2
django-guardian==2.3.0
django-heroku==0.3.1
django-log-viewer==1.0.6
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-s3-storage==0.13.0
django-searchable-encrypted-fields==0.1.4
django-sslify==0.2.7
django-sslserver==0.22
django-storages==1.9.1
django-verification-token==0.0.6
djangorestframework==3.11.0
docopt==0.6.2
docutils==0.15.2
gitignore==0.0.8
gunicorn==20.0.4
idna==2.10
isort==4.3.21
jmespath==0.9.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.1
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.9.8
pylint==2.5.3
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
requests==2.24.0
s3transfer==0.3.3
sentry-sdk==0.16.1
six==1.14.0
sqlparse==0.3.1
static3==0.7.0
stripe==2.48.0
toml==0.10.1
urllib3==1.25.8
whitenoise==5.0.1
wrapt==1.12.1

